fix: tile update bugs fixed
